// packages/domain/src/shared/framework-type.ts
export enum FrameworkType {
  PEST = "PEST",
  FIVE_FORCES = "FIVE_FORCES",
  INTERNAL_ANALYSIS = "INTERNAL_ANALYSIS",
  VRIO = "VRIO",
  THREE_C_PLUS_C = "THREE_C_PLUS_C",
  SWOT = "SWOT",
  CROSS_SWOT = "CROSS_SWOT",
  SEGMENTATION = "SEGMENTATION",
  TARGETING = "TARGETING",
  POSITIONING = "POSITIONING",
  CONCEPT_SHEET = "CONCEPT_SHEET",
  PRODUCT_4P = "PRODUCT_4P",
  PRICE_4P = "PRICE_4P",
  PLACE_4P = "PLACE_4P",
  PROMOTION_4P = "PROMOTION_4P",
  FOUR_C_SEVEN_P = "FOUR_C_SEVEN_P",
  BLUE_OCEAN = "BLUE_OCEAN",
  EXPERIENCE_VALUE = "EXPERIENCE_VALUE",
  VALUE_ADD_METHODS = "VALUE_ADD_METHODS",
  KGI_KSF_KPI = "KGI_KSF_KPI",
  CUSTOMER_JOURNEY = "CUSTOMER_JOURNEY",
  CRM_OVERVIEW = "CRM_OVERVIEW",
  CRM_ANALYSIS = "CRM_ANALYSIS",
}